moved set-expired-to-archived to commands
